refactor: move social list management to GameInterface
